+++ contracts/hub-channels/src/lib.rs
@@ -1,9 +1,10 @@
 pub mod contract;
 mod error;
 pub mod msg;
-pub mod query;
+pub mod execute;
 pub mod state;
 pub mod validating;
 mod tests;
+mod query;
 
 pub use crate::error::ContractError;
